backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.2.2
google-generativeai==0.8.6
yfinance==1.2.0
pandas==2.3.3
ta==0.11.0
newsapi-python==0.2.7
